Inherit from IPythonKernel
The kernel was a basic wrapper kernel, inheriting from KernelBase. By elevating it to extend from the full IPythonKernel, we have a real local ipython kernel running which is now used to run %local cells.

This also gives access to the rich output capture, parse and display engine from ipython, which is now in use.
things like `display()` and the ipywidgets now work correctlt in %local cells.